requirements.txt
streamlit==1.38.0
requests==2.32.3
openai==1.52.2
python-dotenv==1.0.1